LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-redirect-replace-expected.txt


FAIL precommitHandler redirect() replace assert_equals: expected "#redirect2" but got "#replace"
